+++ src/main/java/org/ezcode/codetest/presentation/problemmanagement/testcase/TestCaseController.java
@@ -1,0 +1,4 @@
+package org.ezcode.codetest.presentation.problemmanagement.testcase;
+
+public class TestCaseController {
+}
